perf: subset Roboto + drop dead ConditionalWrapper + tighten heatmap
Roboto subset (TECH-DEBT #4)
- Generalise scripts/subset-monaspace.mjs into subset-fonts.mjs
  driving a list of fonts. Pattern same as before: ASCII letters /
  digits / punctuation + Spanish accents.
- Roboto: 209,484 -> 58,468 bytes (72% smaller). Variable axes
  (wdth, wght) preserved by subset-font, so weights 400/700 still
  render correctly from the same @font-face declaration.
- Bump filename to Roboto-VariableFont_wdth,wght.v2.woff2 to
  cache-bust past the immutable 1y cache header on /fonts/*.
- Roboto preloads on every route, so the 150KB saving propagates
  to every page (home, skills, skills/:uuid, education,
  education/:slug).

Drop ConditionalWrapper (TECH-DEBT #5)
- Component had no real consumers after the NavBar refactor
  removed the only call site (ConditionalLink wrapping a
  constant-URL anchor). Tests + story were the only references
  left.
- Delete the directory, drop the doc reference in AGENTS.md.

Heatmap cell width
- Desktop cells were 14px (GitHub density) but the year header
  labels were squashed at that width and hard to read. Bump to
  18px - tighter than mobile's 22px but legible.

TECH-DEBT.md
- Strike #4 and #5 from the table; remove their sections.

Visual baselines
- Regenerate education-detail (chromium + mobile) and home-mobile
  PNGs. Roboto re-encode shifts text glyph metrics on every route
  that renders English copy at a non-trivial font size.
